requirements.txt
torch
torchvision
Pillow
numpy
jupyter
